creating seeder, factory

+++ database/migrations/2023_06_22_234147_create_books_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->nullable(false);
             $table->decimal('value')->default(0);
-            $table->integer('isbn')->default(0);
+            $table->bigInteger('isbn')->default(0);
             $table->timestamps();
         });
     }
